feat: Add `Invoice No` on POS order.
